+++ openeogeotrellis/deploy/batch_job.py
@@ -292,6 +292,7 @@ def run_job(
 
         backend_implementation = GeoPySparkBackendImplementation(
             use_job_registry=bool(get_backend_config().ejr_api),
+            do_ejr_health_check=False,
         )
 
         if default_sentinel_hub_credentials is not None:
@@ -348,7 +349,7 @@ def run_job(
         global_metadata_attributes = {
             "title": job_specification.get("title", ""),
             "description": job_specification.get("description", ""),
-            "institution": "openEO platform - Geotrellis backend: " + __version__
+            "institution": f"{get_backend_config().processing_facility} - {get_backend_config().capabilities_backend_version}"
         }
 
         ml_model_metadata = None